bridge/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
google-genai>=0.8.0
websockets>=13.0
